pyproject.toml
[project]
name = "disney-queue-agent"
version = "0.1.0"
description = "Disney parks queue analytics agent (Google ADK + Vertex AI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "litellm>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "pandas>=2.2.0",
    "pydantic>=2.0.0",
    "chromadb>=0.6.0",
    "google-cloud-bigquery>=3.41.0",
]
